Implement `MessageDefObjectGenerator`.
