--- codegen-plugins/message-fields/src/main/kotlin/io/spine/chords/protodata/plugin/FileGenerator.kt
@@ -1,108 +1,0 @@
-/*
- * Copyright 2024, TeamDev. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * https://www.apache.org/licenses/LICENSE-2.0
- *
- * Redistribution and use in source and/or binary forms, with or without
- * modification, must retain the above copyright notice and the following
- * disclaimer.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-package io.spine.chords.protodata.plugin
-
-import com.google.protobuf.StringValue
-import io.spine.protodata.ProtoFileHeader
-import io.spine.protodata.TypeName
-import io.spine.protodata.find
-import io.spine.protodata.type.TypeSystem
-import io.spine.string.camelCase
-
-/**
- * Returns a Java package for the [TypeName].
- */
-internal fun TypeName.javaPackage(typeSystem: TypeSystem): String {
-    val messageToHeader = typeSystem.findMessage(this)
-    checkNotNull(messageToHeader) {
-        "Cannot determine file header for TypeName `$this`"
-    }
-    return messageToHeader.second.javaPackage
-}
-
-/**
- * Converts a Proto field name to "property" name,
- * e.g. "domain_name" -> "domainName".
- */
-internal val String.propertyName
-    get() = camelCase().replaceFirstChar { it.lowercase() }
-
-/**
- * Returns a Java package declared in [ProtoFileHeader].
- */
-internal val ProtoFileHeader.javaPackage: String
-    get() {
-        val optionName = "java_package"
-        val option = optionList.find(optionName, StringValue::class.java)
-        checkNotNull(option) {
-            "Cannot find option `$optionName` in file header `${this}`."
-        }
-        return option.value
-    }
-
-/**
- * Generates a simple class name for the implementation of `MessageField`.
- */
-internal fun TypeName.messageFieldClassName(fieldName: String): String {
-    return generateClassName(fieldName, "Field")
-}
-
-/**
- * Generates a simple class name for the implementation of `MessageOneof`.
- */
-internal fun TypeName.messageOneofClassName(fieldName: String): String {
-    return generateClassName(fieldName, "Oneof")
-}
-
-/**
- * Generates a simple class name for the `fieldName` provided.
- */
-internal fun TypeName.generateClassName(suffix: String): String {
-    return generateClassName("", suffix)
-}
-
-/**
- * Generates a simple class name for the [fieldName] provided.
- */
-internal fun TypeName.generateClassName(fieldName: String, suffix: String): String {
-    return nestingTypeNameList.joinToString(
-        "",
-        "",
-        "${simpleName}${fieldName.camelCase()}$suffix"
-    )
-}
-
-/**
- * Returns simple class name for the [TypeName].
- */
-internal val TypeName.simpleClassName: String
-    get() = if (nestingTypeNameList.isNotEmpty())
-        nestingTypeNameList.joinToString(
-            ".",
-            "",
-            ".$simpleName"
-        ) else simpleName
