+++ codegen-plugins/message-fields/src/main/kotlin/io/spine/chords/protodata/plugin/MessageFieldsRenderer.kt
@@ -35,7 +35,7 @@ import io.spine.tools.code.Kotlin
  * implementations for the fields of Proto messages.
  *
  * The renderer prepares all the required data for code generation
- * and runs specific [FileGenerator]s to generate Kotlin files.
+ * and runs specific [MessageDefFileGenerator]s to generate Kotlin files.
  */
 public class MessageFieldsRenderer : Renderer<Kotlin>(Kotlin.lang()) {
 
@@ -69,7 +69,7 @@ public class MessageFieldsRenderer : Renderer<Kotlin>(Kotlin.lang()) {
                     "Message not found for type `$typeName`."
                 }
                 messageToHeader.first.fieldList.let { fields ->
-                    MessageDefGenerator(typeName, fields, typeSystem!!)
+                    MessageDefFileGenerator(typeName, fields, typeSystem!!)
                         .let { generator ->
                             sources.createFile(
                                 generator.filePath(),
